+++ src/main/kotlin/com/atlassian/performance/tools/awsinfrastructure/api/CustomDatasetSource.kt
@@ -1,6 +1,7 @@
 package com.atlassian.performance.tools.awsinfrastructure.api
 
 import com.atlassian.performance.tools.aws.api.StorageLocation
+import com.atlassian.performance.tools.awsinfrastructure.api.jira.StartedNode
 import com.atlassian.performance.tools.concurrency.api.submitWithLogContext
 import com.atlassian.performance.tools.infrastructure.api.dataset.Dataset
 import com.atlassian.performance.tools.ssh.api.Ssh
@@ -17,8 +18,10 @@ class CustomDatasetSource private constructor(
     val jiraHome: RemoteLocation,
     val database: RemoteLocation,
     private val jiraHomeTimeouts: Timeouts,
-    private val databaseTimeouts: Timeouts
+    private val databaseTimeouts: Timeouts,
+    private val nodes: List<StartedNode.StoppableNode>
 ) {
+
     object FileNames {
         const val DATABASE = "database"
         const val JIRAHOME = "jirahome"
@@ -26,36 +29,13 @@ class CustomDatasetSource private constructor(
 
     private val logger: Logger = LogManager.getLogger(this::class.java)
 
-    @Deprecated("Use Builder.CustomDatasetSource instead")
-    constructor(
-        jiraHome: RemoteLocation,
-        database: RemoteLocation
-    ) : this(
-        jiraHome,
-        database,
-        Timeouts(
-            Duration.ofMinutes(25),
-            Duration.ofMinutes(10),
-            Duration.ofMinutes(1)
-        ),
-        Timeouts(
-            Duration.ofMinutes(25),
-            Duration.ofMinutes(10),
-            Duration.ofMinutes(1)
-        )
-    )
-
-    @Suppress("DEPRECATION")
-    @Deprecated("Use Use Builder.CustomDatasetSource.fromJson instead")
-    constructor(json: JsonObject) : this(
-        jiraHome = RemoteLocation(json.getJsonObject("jiraHome")),
-        database = RemoteLocation(json.getJsonObject("database"))
-    )
-
     fun toJson(): JsonObject {
         return Json.createObjectBuilder()
             .add("jiraHome", jiraHome.toJson())
             .add("database", database.toJson())
+            .add("nodes", Json.createArrayBuilder().apply {
+                nodes.map { it.toJson() }.forEach { add(it) }
+            }.build())
             .build()
     }
 
@@ -76,7 +56,7 @@ class CustomDatasetSource private constructor(
                 .build()
         )
 
-        stopJira(jiraHome.host)
+        stopJira()
         stopDockerContainers(database.host)
 
         val jiraHomeUpload = executor.submitWithLogContext("jiraHome") {
@@ -109,10 +89,8 @@ class CustomDatasetSource private constructor(
         return DatasetCatalogue().custom(location)
     }
 
-    private fun stopJira(host: SshHost) {
-        val shutdownTomcat = "echo SHUTDOWN | nc localhost 8005"
-        val waitForNoJavaProcess = "while ps -C java -o pid=; do sleep 5; done"
-        Ssh(host, connectivityPatience = 4).newConnection().use { it.safeExecute("$shutdownTomcat && $waitForNoJavaProcess", Duration.ofMinutes(3)) }
+    private fun stopJira() {
+        nodes.forEach { it.stopNode() }
     }
 
     private fun stopDockerContainers(host: SshHost) {
@@ -131,14 +109,9 @@ class CustomDatasetSource private constructor(
 
     class Builder(
         private var jiraHome: RemoteLocation,
-        private var database: RemoteLocation
+        private var database: RemoteLocation,
+        private var nodes: List<StartedNode.StoppableNode>
     ) {
-
-        constructor(json: JsonObject) : this(
-            jiraHome = RemoteLocation(json.getJsonObject("jiraHome")),
-            database = RemoteLocation(json.getJsonObject("database"))
-        )
-
         private var jiraHomeArchiveTimeout: Duration = Duration.ofMinutes(25)
         private var jiraHomeUploadTimeout: Duration = Duration.ofMinutes(10)
         private var jiraHomeMoveTimeout: Duration = Duration.ofMinutes(1)
@@ -146,14 +119,34 @@ class CustomDatasetSource private constructor(
         private var databaseUploadTimeout: Duration = Duration.ofMinutes(10)
         private var databaseMoveTimeout: Duration = Duration.ofMinutes(1)
 
+        constructor(json: JsonObject) : this(
+            jiraHome = RemoteLocation(json.getJsonObject("jiraHome")),
+            database = RemoteLocation(json.getJsonObject("database")),
+            nodes = json.getJsonArray("nodes").map { StartedNode.StoppableNode.Builder(it.asJsonObject()).build() }
+        )
+
+
         fun jiraHome(jiraHome: RemoteLocation) = apply { this.jiraHome = jiraHome }
         fun database(database: RemoteLocation) = apply { this.database = database }
-        fun jiraHomeArchiveTimeout(jiraHomeArchiveTimeout: Duration) = apply { this.jiraHomeArchiveTimeout = jiraHomeArchiveTimeout }
-        fun jiraHomeUploadTimeout(jiraHomeUploadTimeout: Duration) = apply { this.jiraHomeUploadTimeout = jiraHomeUploadTimeout }
-        fun jiraHomeMoveTimeout(jiraHomeMoveTimeout: Duration) = apply { this.jiraHomeMoveTimeout = jiraHomeMoveTimeout }
-        fun databaseArchiveTimeout(databaseArchiveTimeout: Duration) = apply { this.databaseArchiveTimeout = databaseArchiveTimeout }
-        fun databaseUploadTimeout(databaseUploadTimeout: Duration) = apply { this.databaseUploadTimeout = databaseUploadTimeout }
-        fun databaseMoveTimeout(databaseMoveTimeout: Duration) = apply { this.databaseMoveTimeout = databaseMoveTimeout }
+        fun jiraHomeArchiveTimeout(jiraHomeArchiveTimeout: Duration) =
+            apply { this.jiraHomeArchiveTimeout = jiraHomeArchiveTimeout }
+
+        fun jiraHomeUploadTimeout(jiraHomeUploadTimeout: Duration) =
+            apply { this.jiraHomeUploadTimeout = jiraHomeUploadTimeout }
+
+        fun jiraHomeMoveTimeout(jiraHomeMoveTimeout: Duration) =
+            apply { this.jiraHomeMoveTimeout = jiraHomeMoveTimeout }
+
+        fun databaseArchiveTimeout(databaseArchiveTimeout: Duration) =
+            apply { this.databaseArchiveTimeout = databaseArchiveTimeout }
+
+        fun databaseUploadTimeout(databaseUploadTimeout: Duration) =
+            apply { this.databaseUploadTimeout = databaseUploadTimeout }
+
+        fun databaseMoveTimeout(databaseMoveTimeout: Duration) =
+            apply { this.databaseMoveTimeout = databaseMoveTimeout }
+
+        fun nodes(nodes: List<StartedNode.StoppableNode>) = apply { this.nodes = nodes }
 
         fun build() = CustomDatasetSource(
             jiraHome,
@@ -167,7 +160,8 @@ class CustomDatasetSource private constructor(
                 databaseArchiveTimeout,
                 databaseUploadTimeout,
                 databaseMoveTimeout
-            )
+            ),
+            nodes
         )
     }
 }